python/b35/pyproject.toml
[project]
name = "b35"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-ninja>=1.4.3",
    "django-ninja-jwt>=5.3.7",
    "requests>=2.32.5",
]
